added a new module called R.C.Random.PRG
