+++ Raaz/Core/Random.hs
@@ -22,36 +22,10 @@ import Foreign.Storable(Storable, peek)
 import System.IO ( openBinaryFile, Handle, IOMode(ReadMode)
                  , BufferMode(NoBuffering), hSetBuffering
                  )
-
-
 import Prelude
 
 import Raaz.Core.Types
-
--- | The class that captures pseudo-random generators. Essentially the
--- a pseudo-random generator (PRG) is a byte sources that can be
--- seeded. It is expected that the `fillRandomBytes` call should fill
--- in high quality random bytes in to the Pointer.
-class PRG prg where
-
-  fillRandomBytes :: BYTES Int -> Pointer -> prg -> IO ()
-
-
--- | The class captures prgs that can be seeded.
-class Seedable prg where
-
-  -- | Seed a given prg from a valid source prg.
-  seedFrom :: PRG srcPrg => srcPrg -> prg -> IO ()
-
--- | Generalised version of fillRandomBytes that takes an arbitrary length
--- unit.
-fillRandom :: (PRG prg, LengthUnit l)
-           => l
-           -> Pointer
-           -> prg
-           -> IO ()
-fillRandom = fillRandomBytes . inBytes
-
+import Raaz.Core.Random.PRG
 
 -- | Stuff that can be generated by a pseudo-random generator.
 class Random r where
@@ -83,7 +57,6 @@ instance (Random a, Random b, Random c) => Random (a,b,c) where
   getRandom prg = (,,) <$> getRandom prg <*> getRandom prg <*> getRandom prg
 
 #ifdef HAVE_SYSTEM_PRG
-
 -- | The system wide pseudo-random generator. The source is expected
 -- to be of high quality, albeit a bit slow due to system call
 -- overheads. It is expected that this source is automatically seeded
